requirements.txt
djangorestframework
Django==2.1
pytz==2019.2
